src/capstone/backend/APIGatewayJWT2/src/main/java/com/happiest/APIGatewayJWT2/service/MyUserDetailsService.java
package com.happiest.APIGatewayJWT2.service;

import com.happiest.APIGatewayJWT2.model.UserPrincipal;
import com.happiest.APIGatewayJWT2.model.Users;
import com.happiest.APIGatewayJWT2.repository.UserRepo;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.User;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class MyUserDetailsService implements UserDetailsService {

    @Autowired
    private UserRepo repo;

    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {

        Optional<Users> user = repo.findByEmail(email);

        if (user.isEmpty()) {
            throw new UsernameNotFoundException("User not found with email: " + email);
        }
        return new UserPrincipal(user.get());
    }
}